Moves reader APIs to reader module
