+++ partiql-tools/src/common.rs
@@ -11,7 +11,8 @@ use partiql_catalog::table_fn::{
 use partiql_eval::error::PlanErr;
 use partiql_eval::eval::EvalPlan;
 use partiql_eval::plan::{EvaluationMode, EvaluatorPlanner};
-use partiql_eval::{DataCatalog, ReaderFactory};
+use partiql_eval::reader::ReaderFactory;
+use partiql_eval::DataCatalog;
 use partiql_extension_ion::decode::{IonDecoderBuilder, IonDecoderConfig};
 use partiql_extension_ion::Encoding;
 use partiql_logical::LogicalPlan;
